Convert test_connection to pytest.
